+++ erdpy/projects/project_clang.py
@@ -1,4 +1,3 @@
-import binascii
 import logging
 import subprocess
 from os import path
@@ -57,15 +56,18 @@ class ProjectClang(Project):
     def _do_wasm(self):
         logger.info("_do_wasm")
         tool = path.join(self._get_llvm_path(), "wasm-ld")
+        undefined_file = str(self.build_configuration.undefined_file)
         args = [
             tool,
-            "--verbose",
             "--no-entry",
             str(self.file_o),
             "-o", self.get_file_wasm(),
             "--strip-all",
-            f"-allow-undefined-file={str(self.build_configuration.undefined_file)}"
+            f"-allow-undefined-file={undefined_file}"
         ]
+
+        if self.options.get("verbose", False):
+            args.append("--verbose")
 
         for export in self.build_configuration.exports:
             args.append(f"-export={export}")